linux: Refactor shared-page init.
This is all very confusing and it does not have to be. Simplify things
using aliases and one statement per line.
fb2m is used as two different things depending on the structure it is
defined in, which makes everything worse... That should be changed too.

In the shared page, it is a page-directory, in the xenfb2_info it is the
full translation array (actual fb2m that is)...

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>